Remove legacy admin CLI entrypoints
