+++ tests/test_main_github_ingress.py
@@ -9,7 +9,7 @@ from unittest.mock import patch
 
 from app.github_ingress_runtime import GitHubIssueAssignmentEvent, GitHubPullRequestReviewEvent
 from app.broker import EgressQueueMessage, TaskQueueMessage
-from app.main_message_handler import _load_config, main
+from app.main_message_handler import _load_config, _load_schedule_jobs, main
 from app.models import OutboundMessage, ReplyChannel, TaskEnvelope
 @dataclass
 class _FakeBroker:
@@ -50,6 +50,150 @@ class MainGitHubIngressTests(unittest.TestCase):
 
             with self.assertRaisesRegex(ValueError, "unknown keys"):
                 _load_config(str(config_path))
+
+    def test_load_schedule_jobs_accepts_cron_timezone_and_interval_fallback(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "cron-job",
+                            "content": "Run cron job",
+                            "cron": "0 8 * * *",
+                            "timezone": "Europe/London",
+                            "interval_seconds": 86400,
+                        },
+                        {
+                            "job_name": "interval-job",
+                            "content": "Run interval job",
+                            "interval_seconds": 300,
+                            "start_at": "2026-03-07T00:00:00Z",
+                            "prompt_context": ["Mention overdue alerts first."],
+                        },
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            jobs = _load_schedule_jobs(str(schedule_path))
+
+            self.assertEqual(len(jobs), 2)
+            self.assertEqual(jobs[0].cron, "0 8 * * *")
+            self.assertEqual(jobs[0].timezone_name, "Europe/London")
+            self.assertEqual(jobs[0].interval_seconds, 86400)
+            self.assertEqual(jobs[1].interval_seconds, 300)
+            self.assertEqual(jobs[1].cron, None)
+            self.assertEqual(jobs[1].prompt_context, ["Mention overdue alerts first."])
+
+    def test_load_schedule_jobs_rejects_invalid_prompt_context(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "interval-job",
+                            "content": "Run interval job",
+                            "interval_seconds": 300,
+                            "prompt_context": [""],
+                        }
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            with self.assertRaisesRegex(
+                ValueError,
+                "prompt_context must be a list of non-empty strings",
+            ):
+                _load_schedule_jobs(str(schedule_path))
+
+    def test_load_schedule_jobs_defaults_cron_timezone_to_utc(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "cron-job",
+                            "content": "Run cron job",
+                            "cron": "0 8 * * *",
+                        }
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            jobs = _load_schedule_jobs(str(schedule_path))
+
+            self.assertEqual(len(jobs), 1)
+            self.assertEqual(jobs[0].timezone_name, "UTC")
+
+    def test_load_schedule_jobs_ignores_interval_anchors_when_cron_is_present(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "cron-job",
+                            "content": "Run cron job",
+                            "cron": "0 8 * * *",
+                            "interval_seconds": "legacy-value",
+                            "start_at": 123,
+                        }
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            jobs = _load_schedule_jobs(str(schedule_path))
+
+            self.assertEqual(len(jobs), 1)
+            self.assertEqual(jobs[0].cron, "0 8 * * *")
+            self.assertEqual(jobs[0].interval_seconds, None)
+            self.assertEqual(jobs[0].start_at, None)
+
+    def test_load_schedule_jobs_rejects_invalid_cron_timezone(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "cron-job",
+                            "content": "Run cron job",
+                            "cron": "0 8 * * *",
+                            "timezone": "Not/AZone",
+                        }
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            with self.assertRaisesRegex(ValueError, "invalid timezone"):
+                _load_schedule_jobs(str(schedule_path))
+
+    def test_load_schedule_jobs_rejects_timezone_without_cron(self) -> None:
+        with tempfile.TemporaryDirectory() as tmpdir:
+            schedule_path = Path(tmpdir) / "schedule.json"
+            schedule_path.write_text(
+                json.dumps(
+                    [
+                        {
+                            "job_name": "interval-job",
+                            "content": "Run interval job",
+                            "interval_seconds": 60,
+                            "timezone": "UTC",
+                        }
+                    ]
+                ),
+                encoding="utf-8",
+            )
+
+            with self.assertRaisesRegex(ValueError, "timezone is only valid with cron"):
+                _load_schedule_jobs(str(schedule_path))
 
     def test_main_message_handler_publishes_assignment_event_once_and_uses_checkpoint(self) -> None:
         with tempfile.TemporaryDirectory() as tmpdir:
